Split test_corpus, testx_corpus
